pyproject.toml
[project]
name = "productapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.6",
    "djangorestframework>=3.16.1",
    "drf-yasg>=1.21.11",
    "pillow>=11.3.0",
]
